requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.36
python-dotenv==1.0.1
python-multipart==0.0.20
PyJWT==2.9.0
bcrypt==4.2.0
# Optional local STT fallback for noisy voice capture
faster-whisper==1.1.1
